pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "datapilot"
version = "2.0.0"
description = "AI-powered dbt project auditor with multi-agent intelligence and enterprise integrations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "DataPilot Team" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "networkx>=3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "groq>=0.4.0",
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "flask>=3.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlglot>=23.0",
    "click>=8.0",
    "structlog>=23.0",
    "tenacity>=8.0",
    "httpx>=0.25.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
airflow = ["apache-airflow>=2.7"]
snowflake = ["snowflake-connector-python>=3.0"]
azure = ["azure-storage-blob>=12.0", "azure-identity>=1.0"]
aws = ["boto3>=1.28"]
dbt = ["dbt-core>=1.7", "dbt-duckdb>=1.8"]
kubernetes = ["kubernetes>=28.0"]
powerbi = ["msal>=1.24"]
kafka = ["confluent-kafka>=2.3"]
all = [
    "datapilot[airflow,snowflake,azure,aws,dbt,kubernetes,powerbi,kafka]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
    "pre-commit>=3.0",
]

[project.scripts]
datapilot = "datapilot.cli:main"

[project.urls]
Homepage = "https://github.com/BVB09TS/Data-Pilot"
Repository = "https://github.com/BVB09TS/Data-Pilot"

[tool.hatch.build.targets.wheel]
packages = ["datapilot"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101", "S603", "S607"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
